UN-37-validate

+++ src/main/java/com/unik/unikForma/service/ClasseService.java
@@ -2,18 +2,22 @@ package com.unik.unikForma.service;
 
 import com.unik.unikForma.dto.ClasseDTO;
 import com.unik.unikForma.entity.Classe;
+import com.unik.unikForma.exception.ClasseNotFoundException;
 import com.unik.unikForma.mapper.ClasseMapper;
 import com.unik.unikForma.repository.ClasseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 
+import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
+@Validated
 public class ClasseService {
 
     private final ClasseRepository classeRepository;
@@ -25,7 +29,7 @@ public class ClasseService {
         this.classeMapper = classeMapper;
     }
 
-    public ClasseDTO saveClasse(ClasseDTO classeDTO) {
+    public ClasseDTO saveClasse(@Valid ClasseDTO classeDTO) {
         Classe classe = classeMapper.toEntity(classeDTO);
         Classe savedClasse = classeRepository.save(classe);
         return classeMapper.toDTO(savedClasse);
@@ -37,16 +41,21 @@ public class ClasseService {
     }
 
     public Optional<ClasseDTO> getClasseById(Long id) {
-        return classeRepository.findById(id)
-                .map(classeMapper::toDTO);
+        return Optional.ofNullable(classeRepository.findById(id)
+                .map(classeMapper::toDTO)
+                .orElseThrow(() -> new ClasseNotFoundException(id)));
     }
 
     public void deleteClasse(Long id) {
+        if (!classeRepository.existsById(id)) {
+            throw new ClasseNotFoundException(id);
+        }
         classeRepository.deleteById(id);
     }
 
-    public ClasseDTO updateClasse(Long id, ClasseDTO updatedClasseDTO) {
-        Classe existingClasse = classeRepository.findById(id).orElse(null);
+
+    public ClasseDTO updateClasse(Long id, @Valid ClasseDTO updatedClasseDTO) {
+        Classe existingClasse = classeRepository.findById(id).orElseThrow(() -> new ClasseNotFoundException(id));
 
         classeMapper.updateEntityFromDTO(updatedClasseDTO, existingClasse);
         Classe updatedClasse = classeRepository.save(existingClasse);
